myProject/requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
certifi==2024.2.2
charset-normalizer==2.0.12
Django==3.2.25
djangorestframework==3.15.1
greenlet==2.0.2
idna==3.7
importlib-metadata==4.8.3
importlib-resources==5.4.0
mssql-django==1.4.2
numpy==1.19.5
pandas==1.1.5
pyodbc==4.0.39
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.27.1
responses==0.17.0
six==1.16.0
SQLAlchemy==1.4.52
sqlparse==0.4.4
typing_extensions==4.1.1
urllib3==1.26.18
zipp==3.6.0
